oops

+++ 0x0B-python-object_relational_mapping/model_state.py
@@ -10,5 +10,5 @@ Base = declarative_base()
 class State(Base):
     """State class for sqlalchemy"""
     __tablename__ = 'states'
-    id = Column(Integer(11), nullable=False, primary_key=True)
+    id = Column(Integer, nullable=False, primary_key=True)
     name = Column(String(128), nullable=False)
